+++ src/Ext/Side/Body.cpp
@@ -42,12 +42,11 @@ void SideExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->ToolTip_Background_Opacity.Read(exINI, pSection, "ToolTip.Background.Opacity");
 	this->ToolTip_Background_BlurSize.Read(exINI, pSection, "ToolTip.Background.BlurSize");
 	this->BriefingTheme = pINI->ReadTheme(pSection, "BriefingTheme", this->BriefingTheme);
-	this->SWSidebarBackground_OnPCX.Read(pINI, pSection, "SWSidebarBackground.OnPCX");
-	this->SWSidebarBackground_OffPCX.Read(pINI, pSection, "SWSidebarBackground.OffPCX");
-	this->SWSidebarBackground_TopPCX.Read(pINI, pSection, "SWSidebarBackground.TopPCX");
-	this->SWSidebarBackground_CenterPCX.Read(pINI, pSection, "SWSidebarBackground.CenterPCX");
-	this->SWSidebarBackground_BottomPCX.Read(pINI, pSection, "SWSidebarBackground.BottomPCX");
-
+	this->SuperWeaponSidebar_OnPCX.Read(pINI, pSection, "SuperWeaponSidebar.OnPCX");
+	this->SuperWeaponSidebar_OffPCX.Read(pINI, pSection, "SuperWeaponSidebar.OffPCX");
+	this->SuperWeaponSidebar_TopPCX.Read(pINI, pSection, "SuperWeaponSidebar.TopPCX");
+	this->SuperWeaponSidebar_CenterPCX.Read(pINI, pSection, "SuperWeaponSidebar.CenterPCX");
+	this->SuperWeaponSidebar_BottomPCX.Read(pINI, pSection, "SuperWeaponSidebar.BottomPCX");
 }
 
 // =============================
@@ -77,11 +76,11 @@ void SideExt::ExtData::Serialize(T& Stm)
 		.Process(this->IngameScore_WinTheme)
 		.Process(this->IngameScore_LoseTheme)
 		.Process(this->BriefingTheme)
-		.Process(this->SWSidebarBackground_OnPCX)
-		.Process(this->SWSidebarBackground_OffPCX)
-		.Process(this->SWSidebarBackground_TopPCX)
-		.Process(this->SWSidebarBackground_CenterPCX)
-		.Process(this->SWSidebarBackground_BottomPCX)
+		.Process(this->SuperWeaponSidebar_OnPCX)
+		.Process(this->SuperWeaponSidebar_OffPCX)
+		.Process(this->SuperWeaponSidebar_TopPCX)
+		.Process(this->SuperWeaponSidebar_CenterPCX)
+		.Process(this->SuperWeaponSidebar_BottomPCX)
 		;
 }