+++ src/Ext/Side/Body.h
@@ -36,11 +36,11 @@ public:
 		Nullable<int> ToolTip_Background_Opacity;
 		Nullable<float> ToolTip_Background_BlurSize;
 		Valueable<int> BriefingTheme;
-		PhobosPCXFile SWSidebarBackground_OnPCX;
-		PhobosPCXFile SWSidebarBackground_OffPCX;
-		PhobosPCXFile SWSidebarBackground_TopPCX;
-		PhobosPCXFile SWSidebarBackground_CenterPCX;
-		PhobosPCXFile SWSidebarBackground_BottomPCX;
+		PhobosPCXFile SuperWeaponSidebar_OnPCX;
+		PhobosPCXFile SuperWeaponSidebar_OffPCX;
+		PhobosPCXFile SuperWeaponSidebar_TopPCX;
+		PhobosPCXFile SuperWeaponSidebar_CenterPCX;
+		PhobosPCXFile SuperWeaponSidebar_BottomPCX;
 
 		ExtData(SideClass* OwnerObject) : Extension<SideClass>(OwnerObject)
 			, ArrayIndex { -1 }
@@ -63,11 +63,11 @@ public:
 			, ToolTip_Background_Opacity { }
 			, ToolTip_Background_BlurSize { }
 			, BriefingTheme { -1 }
-			, SWSidebarBackground_OnPCX {}
-			, SWSidebarBackground_OffPCX {}
-			, SWSidebarBackground_TopPCX {}
-			, SWSidebarBackground_CenterPCX {}
-			, SWSidebarBackground_BottomPCX {}
+			, SuperWeaponSidebar_OnPCX {}
+			, SuperWeaponSidebar_OffPCX {}
+			, SuperWeaponSidebar_TopPCX {}
+			, SuperWeaponSidebar_CenterPCX {}
+			, SuperWeaponSidebar_BottomPCX {}
 		{ }
 
 		virtual ~ExtData() = default;