+++ src/Misc/TacticalButtons.cpp
@@ -1,22 +1,17 @@
 #include "TacticalButtons.h"
-#include "PhobosToolTip.h"
 
 #include <GameOptionsClass.h>
-#include <EventClass.h>
 #include <SuperClass.h>
 #include <AircraftClass.h>
 #include <MessageListClass.h>
 #include <TacticalClass.h>
 #include <WWMouseClass.h>
-#include <CCToolTip.h>
 #include <InputManagerClass.h>
 #include <FPSCounter.h>
 #include <AITriggerTypeClass.h>
 #include <JumpjetLocomotionClass.h>
 
-#include <Ext/Side/Body.h>
 #include <Ext/House/Body.h>
-#include <Ext/SWType/Body.h>
 #include <Ext/Scenario/Body.h>
 #include <Utilities/TemplateDef.h>
 
@@ -49,43 +44,7 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 		}
 	}
 
-	if (const int currentCounts = ScenarioExt::Global()->SWButtonData.size()) // Button index 1-11
-	{
-		const int height = DSurface::Composite->GetHeight() - 32;
-
-		if (this->SuperVisible && pMousePosition->X < 65 && pMousePosition->X >= 5) // Button index 1-10 : Super weapons buttons
-		{
-			int checkHight = (height - 48 * currentCounts - 2 * (currentCounts - 1)) / 2;
-
-			for (int i = 0; i < currentCounts; ++i)
-			{
-				if (pMousePosition->Y < checkHight)
-					break;
-
-				checkHight += 48;
-
-				if (pMousePosition->Y < checkHight)
-					return i + 1;
-
-				checkHight += 2;
-			}
-		}
-
-		if (RulesExt::Global()->SWSidebarBackground) // Button index 11 : SW sidebar switch
-		{
-			if (this->SuperVisible ? (pMousePosition->X < 90 && pMousePosition->X >= 80) : (pMousePosition->X < 10 && pMousePosition->X >= 0))
-			{
-				const int checkHight = height / 2;
-
-				if (pMousePosition->Y < checkHight + 25 && pMousePosition->Y >= checkHight - 25)
-					return 11;
-			}
-		}
-	}
-
-	// TODO New buttons
-
-	if (this->HeroVisible) // Button index 61-68 : Heros buttons
+	if (this->HeroVisible) // Button index 1-8 : Heros buttons
 	{
 		auto& vec = ScenarioExt::Global()->OwnedHeros;
 
@@ -105,7 +64,7 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 					checkHight += 48;
 
 					if (pMousePosition->Y < checkHight)
-						return i + 61;
+						return i + 1;
 
 					checkHight += 2;
 				}
@@ -113,7 +72,7 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 		}
 	}
 
-	if (Phobos::Config::SelectedDisplay_Enable) // Button index 71-100 : Select buttons
+	if (Phobos::Config::SelectedDisplay_Enable) // Button index 11-100 : Select buttons
 	{
 		if (const int currentCounts = Math::min(30, static_cast<int>(this->CurrentSelectCameo.size())))
 		{
@@ -125,10 +84,10 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 
 				if (pMousePosition->X < (gap * (currentCounts - 1) + 60))
 				{
-					const int index = currentCounts + 71;
+					const int index = currentCounts + 11;
 					int checkWidth = 0;
 
-					for (int i = 71; i < index; ++i)
+					for (int i = 11; i < index; ++i)
 					{
 						if (this->RecordIndex == i)
 						{
@@ -164,6 +123,8 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 		}
 	}
 
+	// TODO New buttons
+
 	if (this->CheckMouseOverBackground(pMousePosition))
 		return 0; // Button index 0 : Background
 
@@ -172,23 +133,7 @@ int TacticalButtonsClass::CheckMouseOverButtons(const Point2D* pMousePosition)
 
 bool TacticalButtonsClass::CheckMouseOverBackground(const Point2D* pMousePosition)
 {
-	if (RulesExt::Global()->SWSidebarBackground && this->SuperVisible)
-	{
-		if (const int currentCounts = ScenarioExt::Global()->SWButtonData.size())
-		{
-			if (pMousePosition->X < 80 && pMousePosition->X >= 0)
-			{
-				const int height = DSurface::Composite->GetHeight() - 32;
-				const int checkHight = (height - 48 * currentCounts - 2 * (currentCounts - 1)) / 2 - 21;
-
-				if (pMousePosition->Y >= checkHight && pMousePosition->Y < (checkHight + currentCounts * 50 + 40))
-					return true;
-			}
-		}
-	}
-
 	// TODO New button backgrounds
-
 	if (Phobos::Config::SelectedDisplay_Enable)
 	{
 		const int currentCounts = this->CurrentSelectCameo.size();
@@ -239,47 +184,35 @@ void TacticalButtonsClass::SetMouseButtonIndex(const Point2D* pMousePosition)
 {
 	this->ButtonIndex = this->CheckMouseOverButtons(pMousePosition);
 
-	const auto pToolTips = CCToolTip::Instance();
-
-	// SW ToolTip
-	if (this->IndexInSWButtons()) // Button index 1-10 : Super weapons buttons
+	// Hero ToolTip And Button Record
+	if (this->IndexInHerosButtons())
 	{
-		const int superIndex = ScenarioExt::Global()->SWButtonData[this->ButtonIndex - 1];
+		const wchar_t* name = ScenarioExt::Global()->OwnedHeros[this->ButtonIndex - 1]->TypeExtData->OwnerObject()->UIName;
 
-		if (superIndex != -1 && superIndex != this->RecordSuperIndex)
-		{
-			PhobosToolTip::Instance.HelpText_Super(superIndex);
-			this->RecordSuperIndex = superIndex;
-
-			if (pToolTips->ToolTipDelay)
-				pToolTips->LastToolTipDelay = pToolTips->ToolTipDelay;
-
-			pToolTips->ToolTipDelay = 0;
-		}
+		if (!name)
+			this->HoveredHero = nullptr;
+		else if (name != this->HoveredHero)
+			this->HoveredHero = name;
 	}
-	else if (this->RecordSuperIndex != -1)
+	else if (this->HoveredHero)
 	{
-		this->RecordSuperIndex = -1;
-		pToolTips->ToolTipDelay = pToolTips->LastToolTipDelay;
+		this->HoveredHero = nullptr;
 	}
-
-	// TODO New buttons
 
 	// Select ToolTip And Button Record
-	if (this->IndexInSelectButtons()) // Button index 71-100 : Select buttons
+	if (this->IndexInSelectButtons())
 	{
 		this->RecordIndex = this->ButtonIndex;
 		const int currentCounts = this->CurrentSelectCameo.size();
 
 		if (currentCounts > 1 || (currentCounts && this->CurrentSelectCameo[0].Count > 1))
 		{
-			if (TechnoTypeClass* const pType = this->CurrentSelectCameo[this->ButtonIndex - 71].TypeExt->OwnerObject())
-			{
-				const wchar_t* name = pType->UIName;
+			const wchar_t* name = this->CurrentSelectCameo[this->ButtonIndex - 11].TypeExt->OwnerObject()->UIName;
 
-				if (name && name != this->HoveredSelected)
-					this->HoveredSelected = name;
-			}
+			if (!name)
+				this->HoveredSelected = nullptr;
+			else if (name != this->HoveredSelected)
+				this->HoveredSelected = name;
 		}
 		else if (this->HoveredSelected)
 		{
@@ -290,6 +223,8 @@ void TacticalButtonsClass::SetMouseButtonIndex(const Point2D* pMousePosition)
 	{
 		this->HoveredSelected = nullptr;
 	}
+
+	// TODO New buttons
 }
 
 void TacticalButtonsClass::PressDesignatedButton(int triggerIndex)
@@ -297,27 +232,7 @@ void TacticalButtonsClass::PressDesignatedButton(int triggerIndex)
 	if (!this->MouseIsOverButtons()) // In buttons background
 		return;
 
-	if (this->IndexInSWButtons()) // Button index 1-10 : Super weapons buttons
-	{
-		if (!triggerIndex)
-			this->SWSidebarTrigger(this->ButtonIndex);
-		else if (triggerIndex == 2)
-			DisplayClass::Instance->CurrentSWTypeIndex = -1;
-
-		const auto pToolTips = CCToolTip::Instance();
-		this->RecordSuperIndex = -1;
-		pToolTips->ToolTipDelay = pToolTips->LastToolTipDelay;
-	}
-	else if (this->IndexIsSWSwitch())
-	{
-		if (!triggerIndex)
-			this->SWSidebarSwitch();
-	}
-/*	else if (?) // TODO New buttons
-	{
-		;
-	}*/
-	else if (this->IndexInHerosButtons())
+	if (this->IndexInHerosButtons())
 	{
 		if (!triggerIndex)
 			this->HeroSelect(this->ButtonIndex);
@@ -329,6 +244,7 @@ void TacticalButtonsClass::PressDesignatedButton(int triggerIndex)
 		else if (triggerIndex == 2)
 			this->SelectedTrigger(this->ButtonIndex, false);
 	}
+	// TODO New buttons
 }
 
 #pragma endregion
@@ -713,7 +629,8 @@ void TacticalButtonsClass::CurrentSelectInfoDraw()
 			else
 				drawText("PathDir = %d , %d , %d , %d , %d , %d , %d , %d", pD[0], pD[1], pD[2], pD[3], pD[4], pD[5], pD[6], pD[7]);
 
-			drawText("SpeedPercentage = %d", static_cast<int>(pFoot->SpeedPercentage * 100));
+			drawText("CurrentSpeed = %d", static_cast<int>(pFoot->GetCurrentSpeed()));
+			drawText("PercentSpeed = %d", static_cast<int>(pFoot->SpeedPercentage * 100));
 
 			drawInfo("Destination", pFoot, pFoot->Destination);
 			drawInfo("Last Destination", pFoot, pFoot->LastDestination);
@@ -903,360 +820,12 @@ void TacticalButtonsClass::CurrentSelectInfoDraw()
 
 #pragma endregion
 
-#pragma region SWSidebarButtons
+#pragma region HerosButtons
 
-inline bool TacticalButtonsClass::IndexInSWButtons()
+inline bool TacticalButtonsClass::IndexInHerosButtons()
 {
-	return this->ButtonIndex > 0 && this->ButtonIndex <= 10;
+	return this->ButtonIndex > 0 && this->ButtonIndex <= 8;
 }
-
-void TacticalButtonsClass::SWSidebarDraw()
-{
-	auto& data = ScenarioExt::Global()->SWButtonData;
-	const int currentCounts = data.size();
-
-	if (!currentCounts)
-		return;
-
-	const auto pHouse = HouseClass::CurrentPlayer();
-	const auto pSideExt = SideExt::ExtMap.Find(SideClass::Array->GetItem(pHouse->SideIndex));
-	const bool drawSWSidebarBackground = RulesExt::Global()->SWSidebarBackground && pSideExt;
-	const int height = DSurface::Composite->GetHeight() - 32;
-	const int color = Drawing::RGB_To_Int(Drawing::TooltipColor);
-
-	// Draw switch
-	if (drawSWSidebarBackground)
-	{
-		RectangleStruct drawRect { (this->SuperVisible ? 80 : 0), (height / 2 - 25), 10, 50 };
-
-		if (const auto CameoPCX = (this->SuperVisible ? pSideExt->SWSidebarBackground_OnPCX.GetSurface() : pSideExt->SWSidebarBackground_OffPCX.GetSurface()))
-			PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, CameoPCX);
-		else
-			DSurface::Composite->FillRect(&drawRect, COLOR_BLUE);
-
-		if (this->IndexIsSWSwitch())
-		{
-			RectangleStruct rect { 0, 0, (this->SuperVisible ? 90 : 10), drawRect.Y + 50 };
-			DSurface::Composite->DrawRectEx(&rect, &drawRect, color);
-		}
-	}
-
-	if (!this->SuperVisible)
-		return;
-
-	Point2D position { 5, (height - 48 * currentCounts - 2 * (currentCounts - 1)) / 2 };
-	RectangleStruct rect { 0, 0, 65, position.Y + 48 };
-	int recordHeight = -1;
-
-	// Draw top background (80 * 20)
-	Point2D backPosition { 0, position.Y - 21 };
-
-	if (drawSWSidebarBackground)
-	{
-		RectangleStruct drawRect { 0, backPosition.Y, 80, 20 };
-
-		if (const auto CameoPCX = pSideExt->SWSidebarBackground_TopPCX.GetSurface())
-			PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, CameoPCX);
-		else
-			DSurface::Composite->FillRect(&drawRect, COLOR_BLACK);
-	}
-
-	// Draw each buttons
-	for (int i = 0; i < currentCounts; position.Y += 50, rect.Height += 50) // Button index 1-10
-	{
-		// Draw center background (80 * 50)
-		if (drawSWSidebarBackground)
-		{
-			backPosition.Y = position.Y - 1;
-			RectangleStruct drawRect { 0, backPosition.Y, 80, 50 };
-
-			if (const auto CameoPCX = pSideExt->SWSidebarBackground_CenterPCX.GetSurface())
-				PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, CameoPCX);
-			else
-				DSurface::Composite->FillRect(&drawRect, COLOR_BLACK);
-		}
-
-		// Get SW data
-		const auto pSuper = pHouse->Supers.Items[data[i]];
-		const auto pSWType = pSuper->Type;
-		bool canAfford = true;
-
-		// Draw cameo
-		if (const auto pTypeExt = SWTypeExt::ExtMap.Find(pSWType))
-		{
-			if (const auto CameoPCX = pTypeExt->SidebarPCX.GetSurface())
-			{
-				RectangleStruct drawRect { position.X, position.Y, 60, 48 };
-				PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, CameoPCX);
-			}
-			else if (const auto pSHP = pSWType->SidebarImage)
-			{
-				if (const auto MissingCameoPCX = this->GetMissingCameo(pSHP))
-				{
-					RectangleStruct drawRect { position.X, position.Y, 60, 48 };
-					PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, MissingCameoPCX);
-				}
-				else
-				{
-					DSurface::Composite->DrawSHP(pTypeExt->CameoPal.GetOrDefaultConvert(FileSystem::CAMEO_PAL), pSHP, 0, &position, &rect,
-						BlitterFlags::bf_400, 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
-				}
-			}
-
-			if (!pHouse->CanTransactMoney(pTypeExt->Money_Amount))
-				canAfford = false;
-		}
-
-		const bool ready = !pSuper->IsSuspended && (pSWType->UseChargeDrain ? pSuper->ChargeDrainState == ChargeDrainState::Ready : pSuper->IsReady);
-
-		// Flash cameo
-		if (ready && canAfford)
-		{
-			const int delay = pSWType->FlashSidebarTabFrames;
-
-			if (delay > 0 && ((Unsorted::CurrentFrame - pSuper->ReadyFrame) % (delay << 1)) > delay)
-			{
-				DSurface::Composite->DrawSHP(FileSystem::SIDEBAR_PAL, Make_Global<SHPStruct*>(0xB07BC0), 0, &position, &rect,
-					(BlitterFlags::bf_400 | BlitterFlags::TransLucent75), 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
-			}
-		}
-
-		// SW charge progress
-		if (pSuper->ShouldDrawProgress())
-		{
-			const int frame = pSuper->AnimStage() + 1;
-
-			DSurface::Composite->DrawSHP(FileSystem::SIDEBAR_PAL, Make_Global<SHPStruct*>(0xB0B484), frame, &position, &rect,
-				BlitterFlags(0x404), 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
-		}
-		else if (!canAfford)
-		{
-			DSurface::Composite->DrawSHP(FileSystem::SIDEBAR_PAL, Make_Global<SHPStruct*>(0xB07BC0), 0, &position, &rect,
-				(BlitterFlags::bf_400 | BlitterFlags::Darken), 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
-		}
-
-		// SW status
-		if (ready && !this->KeyCodeText[i].empty())
-		{
-			const auto pKey = this->KeyCodeText[i].c_str();
-			Point2D textLocation { 35, position.Y + 1 };
-			constexpr TextPrintType printType = TextPrintType::Center | TextPrintType::FullShadow | TextPrintType::Point8;
-			auto textRect = Drawing::GetTextDimensions(pKey, textLocation, static_cast<WORD>(printType), 2, 1);
-
-			// Text black background
-			reinterpret_cast<void(__fastcall*)(RectangleStruct*, DSurface*, unsigned short, unsigned char)>(0x621B80)(&textRect, DSurface::Composite, 0, 0xAFu);
-			DSurface::Composite->DrawTextA(pKey, &rect, &textLocation, static_cast<COLORREF>(color), COLOR_BLACK, printType);
-		}
-		else if (const wchar_t* pName = pSuper->NameReadiness())
-		{
-			Point2D textLocation { 35, position.Y + 1 };
-			constexpr TextPrintType printType = TextPrintType::Center | TextPrintType::FullShadow | TextPrintType::Point8;
-			auto textRect = Drawing::GetTextDimensions(pName, textLocation, static_cast<WORD>(printType), 2, 1);
-
-			// Text black background
-			reinterpret_cast<void(__fastcall*)(RectangleStruct*, DSurface*, unsigned short, unsigned char)>(0x621B80)(&textRect, DSurface::Composite, 0, 0xAFu);
-			DSurface::Composite->DrawTextA(pName, &rect, &textLocation, static_cast<COLORREF>(color), COLOR_BLACK, printType);
-		}
-
-		if (++i == this->ButtonIndex)
-			recordHeight = position.Y;
-	}
-
-	// Draw bottom background (80 * 20)
-	if (drawSWSidebarBackground)
-	{
-		backPosition.Y = position.Y - 1;
-		RectangleStruct drawRect { 0, backPosition.Y, 80, 20};
-
-		if (const auto CameoPCX = pSideExt->SWSidebarBackground_BottomPCX.GetSurface())
-			PCX::Instance->BlitToSurface(&drawRect, DSurface::Composite, CameoPCX);
-		else
-			DSurface::Composite->FillRect(&drawRect, COLOR_BLACK);
-	}
-
-	// Draw mouse hover rectangle
-	if (!ScenarioClass::Instance->UserInputLocked && recordHeight >= 0)
-	{
-		position.Y = recordHeight;
-		rect.Height = recordHeight + 48;
-
-		RectangleStruct drawRect { 5, position.Y, 60, 48 };
-		DSurface::Composite->DrawRectEx(&rect, &drawRect, color);
-	}
-}
-
-void TacticalButtonsClass::SWSidebarRecheck()
-{
-	const auto pHouse = HouseClass::CurrentPlayer();
-	auto& data = ScenarioExt::Global()->SWButtonData;
-
-	for (auto it = data.begin(); it != data.end();)
-	{
-		const int superIndex = *it;
-
-		if (superIndex >= pHouse->Supers.Count || !pHouse->Supers.Items[superIndex]->IsPresent)
-			it = data.erase(it);
-		else
-			++it;
-	}
-}
-
-bool TacticalButtonsClass::SWSidebarAdd(int& superIndex)
-{
-	const auto pType = SuperWeaponTypeClass::Array->Items[superIndex];
-	const auto pTypeExt = SWTypeExt::ExtMap.Find(pType);
-	bool overflow = true;
-
-	if (pTypeExt && pTypeExt->SW_InScreen_Show)
-	{
-		const unsigned int ownerBits = 1u << HouseClass::CurrentPlayer->Type->ArrayIndex;
-
-		if (pTypeExt->SW_InScreen_RequiredHouses & ownerBits)
-		{
-			auto& data = ScenarioExt::Global()->SWButtonData;
-			const int currentCounts = data.size();
-			bool move = false;
-
-			for (int i = 0; i < 10; ++i) // 10 buttons at max
-			{
-				if (move)
-				{
-					if (i < currentCounts)
-					{
-						int buffer = data[i];
-						data[i] = superIndex;
-						superIndex = buffer;
-					}
-					else
-					{
-						data.push_back(superIndex);
-						overflow = false;
-						break;
-					}
-				}
-				else if (i < currentCounts)
-				{
-					if (this->SWSidebarSort(SuperWeaponTypeClass::Array->Items[data[i]], pType, pTypeExt, ownerBits))
-					{
-						move = true;
-						int buffer = data[i];
-						data[i] = superIndex;
-						superIndex = buffer;
-					}
-				}
-				else
-				{
-					data.push_back(superIndex);
-					overflow = false;
-					break;
-				}
-			}
-		}
-	}
-
-	return overflow;
-}
-
-bool TacticalButtonsClass::SWSidebarSort(SuperWeaponTypeClass* pDataType, SuperWeaponTypeClass* pAddType, SWTypeExt::ExtData* pAddTypeExt, unsigned int ownerBits)
-{
-	const auto pDataTypeExt = SWTypeExt::ExtMap.Find(pDataType);
-
-	if (pDataTypeExt && pAddTypeExt)
-	{
-		if ((pDataTypeExt->SW_InScreen_PriorityHouses & ownerBits) && !(pAddTypeExt->SW_InScreen_PriorityHouses & ownerBits))
-			return false;
-
-		if (!(pDataTypeExt->SW_InScreen_PriorityHouses & ownerBits) && (pAddTypeExt->SW_InScreen_PriorityHouses & ownerBits))
-			return true;
-
-		if (pDataTypeExt->CameoPriority > pAddTypeExt->CameoPriority)
-			return false;
-
-		if (pDataTypeExt->CameoPriority < pAddTypeExt->CameoPriority)
-			return true;
-	}
-
-	if (pDataType->RechargeTime < pAddType->RechargeTime)
-		return false;
-
-	if (pDataType->RechargeTime > pAddType->RechargeTime)
-		return true;
-
-	return wcscmp(pDataType->UIName, pAddType->UIName) > 0;
-}
-
-void TacticalButtonsClass::SWSidebarTrigger(int buttonIndex)
-{
-	if (ScenarioClass::Instance->UserInputLocked || !this->SuperVisible || HouseClass::CurrentPlayer->IsObserver())
-		return;
-
-	auto& data = ScenarioExt::Global()->SWButtonData;
-
-	if (static_cast<size_t>(buttonIndex) > data.size())
-		return;
-
-	const auto pSidebar = SidebarClass::Instance();
-	this->DummyAction = true;
-
-	DummySelectClass pButton;
-	pButton.LinkTo = &pSidebar->Tabs[pSidebar->ActiveTabIndex];
-	pButton.unknown_int_30 = 0x7FFFFFFF - (2 * pButton.LinkTo->TopRowIndex);
-	pButton.SWIndex = data[buttonIndex - 1];
-
-	DWORD KeyNum = 0;
-	reinterpret_cast<bool(__thiscall*)(DummySelectClass*, GadgetFlag, DWORD*, KeyModifier)>(0x6AAD00)(&pButton, GadgetFlag::LeftPress, &KeyNum, KeyModifier::None); // SelectClass_Action
-}
-
-void TacticalButtonsClass::SWSidebarRecord(int buttonIndex, int key)
-{
-	const int index = buttonIndex - 1;
-
-	if (this->KeyCodeData[index] == key)
-		return;
-
-	this->KeyCodeData[index] = key;
-
-	wchar_t text[0x40];
-	reinterpret_cast<void(__fastcall*)(short, wchar_t*)>(0x61EF70)(static_cast<short>(key), text);
-
-	this->KeyCodeText[index] = text;
-}
-
-#pragma endregion
-
-#pragma region SWSidebarSwitch
-
-inline bool TacticalButtonsClass::IndexIsSWSwitch()
-{
-	return this->ButtonIndex == 11;
-}
-
-void TacticalButtonsClass::SWSidebarSwitch()
-{
-	if (ScenarioClass::Instance->UserInputLocked)
-		return;
-
-	const auto pHouse = HouseClass::CurrentPlayer();
-
-	if (pHouse->IsObserver())
-		return;
-
-	this->SuperVisible = !this->SuperVisible;
-
-	MessageListClass::Instance->PrintMessage
-	(
-		(this->SuperVisible ?
-			GeneralUtils::LoadStringUnlessMissing("TXT_EX_SW_BAR_VISIBLE", L"Set exclusive SW sidebar visible.") :
-			GeneralUtils::LoadStringUnlessMissing("TXT_EX_SW_BAR_INVISIBLE", L"Set exclusive SW sidebar invisible.")),
-		RulesClass::Instance->MessageDelay,
-		pHouse->ColorSchemeIndex,
-		true
-	);
-}
-
-#pragma endregion
-
-#pragma region SWExtraFunctions
 
 BSurface* TacticalButtonsClass::GetMissingCameo(SHPStruct* pSHP)
 {
@@ -1276,70 +845,6 @@ BSurface* TacticalButtonsClass::GetMissingCameo(SHPStruct* pSHP)
 	return nullptr;
 }
 
-bool TacticalButtonsClass::SWQuickLaunch(int superIndex)
-{
-	bool keyboardCall = false;
-
-	if (this->KeyboardCall)
-	{
-		this->KeyboardCall = false;
-		keyboardCall = true;
-	}
-
-	const auto pType = SuperWeaponTypeClass::Array->Items[superIndex];
-
-	if (const auto pTypeExt = SWTypeExt::ExtMap.Find(pType))
-	{
-		const int index = HouseClass::CurrentPlayer->ArrayIndex;
-		const auto pTactical = TacticalClass::Instance();
-
-		if (pTypeExt->SW_QuickFireAtMouse && keyboardCall)
-		{
-			const auto mouseCoords = pTactical->ClientToCoords(WWMouseClass::Instance->XY1);
-
-			if (mouseCoords != CoordStruct::Empty)
-			{
-				const EventClass event
-				(
-					index,
-					EventType::SpecialPlace,
-					pType->ArrayIndex,
-					CellClass::Coord2Cell(mouseCoords)
-				);
-				EventClass::AddEvent(event);
-
-				return true;
-			}
-		}
-		else if (!pTypeExt->SW_QuickFireInScreen)
-		{
-			return false;
-		}
-
-		const EventClass event
-		(
-			index,
-			EventType::SpecialPlace,
-			pType->ArrayIndex,
-			CellClass::Coord2Cell(pTactical->ClientToCoords(Point2D{ (DSurface::Composite->Width >> 1), (DSurface::Composite->Height >> 1) }))
-		);
-		EventClass::AddEvent(event);
-
-		return true;
-	}
-
-	return false;
-}
-
-#pragma endregion
-
-#pragma region HerosButtons
-
-inline bool TacticalButtonsClass::IndexInHerosButtons()
-{
-	return this->ButtonIndex > 60 && this->ButtonIndex <= 68;
-}
-
 void TacticalButtonsClass::HerosDraw()
 {
 	if (!this->HeroVisible || HouseClass::CurrentPlayer->IsObserver())
@@ -1353,7 +858,7 @@ void TacticalButtonsClass::HerosDraw()
 
 	Point2D position { DSurface::Composite->GetWidth() - 65, 35 };
 	RectangleStruct drawRect { position.X, position.Y, 60, 48 };
-	const int recordIndex = this->ButtonIndex - 61;
+	const int recordIndex = this->ButtonIndex - 1;
 	int counts = Math::min(8, size);
 
 	for (int i = 0; i < counts; ++i, position.Y += 50, drawRect.Y = position.Y)
@@ -1676,7 +1181,7 @@ void TacticalButtonsClass::HeroSelect(int buttonIndex)
 		return;
 
 	auto& vec = ScenarioExt::Global()->OwnedHeros;
-	const int index = buttonIndex - 61;
+	const int index = buttonIndex - 1;
 
 	if (index >= static_cast<int>(vec.size()))
 		return;
@@ -1693,7 +1198,7 @@ void TacticalButtonsClass::HeroSelect(int buttonIndex)
 		if (ObjectClass::CurrentObjects->Count != 1 || !pSelect->IsSelected)
 			MapClass::UnselectAll();
 
-		if (!pSelect->Select())
+		if (!pSelect->InLimbo && !pSelect->Select())
 			TacticalClass::Instance->SetTacticalPosition(&pSelect->Location);
 	}
 }
@@ -1719,7 +1224,7 @@ void TacticalButtonsClass::HeroSwitch()
 
 inline bool TacticalButtonsClass::IndexInSelectButtons()
 {
-	return this->ButtonIndex > 70 && this->ButtonIndex <= 100;
+	return this->ButtonIndex > 10 && this->ButtonIndex <= 100;
 }
 
 inline void TacticalButtonsClass::AddToCurrentSelect(TechnoTypeExt::ExtData* pTypeExt, int count, int checkIndex)
@@ -1790,10 +1295,10 @@ void TacticalButtonsClass::SelectedTrigger(int buttonIndex, bool select)
 
 	const int currentCounts = this->CurrentSelectCameo.size();
 
-	if (buttonIndex > (currentCounts + 70))
+	if (buttonIndex > (currentCounts + 10))
 		return;
 
-	const auto pTypeExt = this->CurrentSelectCameo[buttonIndex - 71].TypeExt;
+	const auto pTypeExt = this->CurrentSelectCameo[buttonIndex - 11].TypeExt;
 	const auto groupID = pTypeExt->GetSelectionGroupID();
 	const auto shiftPressed = InputManagerClass::Instance->IsForceSelectKeyPressed();
 
@@ -1902,7 +1407,7 @@ void TacticalButtonsClass::SelectedUpdate()
 
 	if (ObjectClass::CurrentObjects->Count <= 0)
 	{
-		this->RecordIndex = 71;
+		this->RecordIndex = 11;
 		return;
 	}
 
@@ -1969,7 +1474,7 @@ void TacticalButtonsClass::SelectedUpdate()
 		}
 	}
 
-	const int maxIndex = this->CurrentSelectCameo.size() + 70;
+	const int maxIndex = this->CurrentSelectCameo.size() + 10;
 
 	if (this->RecordIndex > maxIndex)
 		this->RecordIndex = maxIndex;
@@ -1994,13 +1499,13 @@ void TacticalButtonsClass::SelectedDraw()
 		const int gap = Math::min(60, ((DSurface::Composite->GetWidth() << 1) / (5 * currentCounts)));
 		RectangleStruct surfaceRect { 0, 0, (gap * (currentCounts - 1) + 60), (position.Y + 48) };
 
-		const int maxIndex = currentCounts + 70;
+		const int maxIndex = currentCounts + 10;
 		const COLORREF color = Drawing::RGB_To_Int(Drawing::TooltipColor);
 		TextPrintType printType = TextPrintType::Background | TextPrintType::FullShadow | TextPrintType::Point8;
 
-		for (int i = 71; i < this->RecordIndex && i <= maxIndex; ++i, position.X += gap, drawRect.X = position.X, textPosition.X += gap)
+		for (int i = 11; i < this->RecordIndex && i <= maxIndex; ++i, position.X += gap, drawRect.X = position.X, textPosition.X += gap)
 		{
-			const auto pSelect = this->CurrentSelectCameo[i - 71];
+			const auto pSelect = this->CurrentSelectCameo[i - 11];
 			const auto pTypeExt = pSelect.TypeExt;
 
 			if (const auto CameoPCX = pTypeExt->CameoPCX.GetSurface())
@@ -2039,7 +1544,7 @@ void TacticalButtonsClass::SelectedDraw()
 
 		for (int i = maxIndex; i > this->RecordIndex; --i, position.X -= gap, drawRect.X = position.X, textPosition.X -= gap)
 		{
-			const auto pSelect = this->CurrentSelectCameo[i - 71];
+			const auto pSelect = this->CurrentSelectCameo[i - 11];
 			const auto pTypeExt = pSelect.TypeExt;
 
 			if (const auto CameoPCX = pTypeExt->CameoPCX.GetSurface())
@@ -2085,7 +1590,7 @@ void TacticalButtonsClass::SelectedDraw()
 
 		if (this->RecordIndex <= maxIndex)
 		{
-			const auto pSelect = this->CurrentSelectCameo[this->RecordIndex - 71];
+			const auto pSelect = this->CurrentSelectCameo[this->RecordIndex - 11];
 			const auto pTypeExt = pSelect.TypeExt;
 
 			if (const auto CameoPCX = pTypeExt->CameoPCX.GetSurface())
@@ -2389,7 +1894,7 @@ void TacticalButtonsClass::SelectedSwitch()
 	}
 	else
 	{
-		this->RecordIndex = 71;
+		this->RecordIndex = 11;
 		MessageListClass::Instance->PrintMessage
 		(
 			GeneralUtils::LoadStringUnlessMissing("TXT_SELECT_INVISIBLE", L"Set select info invisible."),
@@ -2529,79 +2034,9 @@ DEFINE_HOOK(0x6D4941, TacticalClass_Render_DrawButtonCameo, 0x6)
 
 	pButtons->SelectedDraw();
 	pButtons->HerosDraw();
-	pButtons->SWSidebarDraw();
 	pButtons->CurrentSelectInfoDraw();
 
 	return 0;
-}
-
-#pragma endregion
-
-#pragma region SWButtonsUpdateHooks
-
-DEFINE_HOOK(0x4F9283, HouseClass_Update_RecheckTechTree, 0x5)
-{
-	GET(HouseClass*, pHouse, ESI);
-
-	if (pHouse == HouseClass::CurrentPlayer)
-		TacticalButtonsClass::Instance.SWSidebarRecheck();
-
-	return 0;
-}
-
-DEFINE_HOOK(0x6A6314, SidebarClass_AddCameo_SupportSWButtons, 0x8)
-{
-	enum { SkipThisCameo = 0x6A65F5 };
-
-	GET_STACK(const AbstractType, absType, STACK_OFFSET(0x14, 0x4));
-	REF_STACK(int, index, STACK_OFFSET(0x14, 0x8));
-
-	return (absType != AbstractType::Special || SuperWeaponTypeClass::Array->Count <= index || TacticalButtonsClass::Instance.SWSidebarAdd(index)) ? 0 : SkipThisCameo;
-}
-
-#pragma endregion
-
-#pragma region SWExtraFunctionHooks
-
-DEFINE_HOOK(0x6AAF46, SelectClass_Action_ButtonClick1, 0x6)
-{
-	enum { SkipClearMouse = 0x6AB95A };
-
-	GET(const int, index, ESI);
-
-	return TacticalButtonsClass::Instance.SWQuickLaunch(index) ? SkipClearMouse : 0;
-}
-
-#pragma endregion
-
-#pragma region SWButtonsTriggerHooks
-
-DEFINE_HOOK_AGAIN(0x6AAD2F, SelectClass_Action_ButtonClick2, 0x7)
-DEFINE_HOOK(0x6AB94F, SelectClass_Action_ButtonClick2, 0xB)
-{
-	enum { ForceEffective = 0x6AAE7C };
-
-	if (!TacticalButtonsClass::Instance.DummyAction)
-		return 0;
-
-	GET(TacticalButtonsClass::DummySelectClass* const , pThis, EDI);
-
-	R->Stack(STACK_OFFSET(0xAC, -0x98), pThis->SWIndex);
-	return ForceEffective;
-}
-
-DEFINE_HOOK(0x6AB961, SelectClass_Action_ButtonClick3, 0x7)
-{
-	enum { SkipControlAction = 0x6AB975 };
-
-	const auto pButtons = &TacticalButtonsClass::Instance;
-
-	if (!pButtons->DummyAction)
-		return 0;
-
-	pButtons->DummyAction = false;
-
-	return SkipControlAction;
 }
 
 #pragma endregion
@@ -2671,46 +2106,29 @@ DEFINE_HOOK(0x5F44FC, ObjectClass_Deselect, 0x7)
 
 #pragma endregion
 
-#pragma region SWShortcutsKeysHooks
+#pragma region ToolTips
 
-DEFINE_HOOK(0x533E69, UnknownClass_sub_533D20_LoadKeyboardCodeFromINI, 0x6)
+DEFINE_HOOK(0x4AE511, DisplayClass_GetToolTip_SkipTacticalTip, 0x5)
 {
-	GET(CommandClass*, pCommand, ESI);
-	GET(int, key, EDI);
+	enum { UseButtonTip = 0x4AE5F8, SkipGameCode = 0x4AE69B };
 
 	const auto pButtons = &TacticalButtonsClass::Instance;
-	const auto name = pCommand->GetName();
-	char buffer[29];
+	const auto buttonIndex = pButtons->GetButtonIndex();
 
-	for (int i = 1; i <= 10; ++i)
-	{
-		sprintf_s(buffer, "SW Sidebar Shortcuts Num %02d", i);
+	if (buttonIndex < 0)
+		return 0;
 
-		if (!_strcmpi(name, buffer))
-			pButtons->SWSidebarRecord(i, key);
-	}
+	if (!buttonIndex)
+		return SkipGameCode;
 
-	return 0;
-}
+	if (pButtons->IndexInHerosButtons())
+		R->EAX(pButtons->HoveredHero);
+	else if (pButtons->IndexInSelectButtons())
+		R->EAX(pButtons->HoveredSelected);
+	else
+		R->EAX(0);
 
-DEFINE_HOOK(0x5FB992, UnknownClass_sub_5FB320_SaveKeyboardCodeToINI, 0x6)
-{
-	GET(CommandClass*, pCommand, ECX);
-	GET(int, key, EAX);
-
-	const auto pButtons = &TacticalButtonsClass::Instance;
-	const auto name = pCommand->GetName();
-	char buffer[30];
-
-	for (int i = 1; i <= 10; ++i)
-	{
-		sprintf_s(buffer, "SW Sidebar Shortcuts Num %02d", i);
-
-		if (!_strcmpi(name, buffer))
-			pButtons->SWSidebarRecord(i, key);
-	}
-
-	return 0;
+	return UseButtonTip;
 }
 
 #pragma endregion