+++ src/Misc/TacticalButtons.h
@@ -30,41 +30,14 @@ public:
 	void CurrentSelectPathDraw();
 	void CurrentSelectInfoDraw();
 
-	// Button index 1-10 : Super weapons buttons
-	inline bool IndexInSWButtons();
-	void SWSidebarDraw();
-	void SWSidebarRecheck();
-	bool SWSidebarAdd(int& superIndex);
-	bool SWSidebarSort(SuperWeaponTypeClass* pDataType, SuperWeaponTypeClass* pAddType, SWTypeExt::ExtData* pAddTypeExt, unsigned int ownerBits);
-	void SWSidebarTrigger(int buttonIndex);
-	void SWSidebarRecord(int buttonIndex, int key);
-
-	struct DummySelectClass
-	{
-		char _[0x2C] {}; // : ControlClass
-		StripClass *LinkTo { nullptr };
-		int unknown_int_30 { 0 };
-		bool MouseEntered { false };
-		int SWIndex { -1 }; // New
-	};
-
-	// Button index 11 : SW sidebar switch
-	inline bool IndexIsSWSwitch();
-	void SWSidebarSwitch();
-
-	// Extra functions for SW
-	BSurface* GetMissingCameo(SHPStruct* pSHP);
-	bool SWQuickLaunch(int superIndex);
-
-	// TODO New buttons
-
-	// Button index 61-68 : Heros buttons
+	// Button index 1-8 : Heros buttons
 	inline bool IndexInHerosButtons();
+	BSurface* GetMissingCameo(SHPStruct* pSHP);
 	void HerosDraw();
 	void HeroSelect(int buttonIndex);
 	void HeroSwitch();
 
-	// Button index 71-100 : Select buttons
+	// Button index 11-100 : Select buttons
 	inline bool IndexInSelectButtons();
 	inline void AddToCurrentSelect(TechnoTypeExt::ExtData* pTypeExt, int count, int checkIndex);
 	BSurface* SearchMissingCameo(AbstractType absType, SHPStruct* pSHP);
@@ -79,6 +52,8 @@ public:
 		int Count { 0 };
 	};
 
+	// TODO New buttons
+
 public:
 	bool PressedInButtonsLayer { false }; // Check press
 
@@ -89,26 +64,17 @@ public:
 
 	// Button index N/A : Show Current Info
 
-	// Button index 1-10 : Super weapons buttons
-	bool DummyAction { false };
-	bool KeyboardCall { false };
-	int RecordSuperIndex { -1 };
-	std::wstring KeyCodeText[10] {};
-	int KeyCodeData[10] {};
-
-	// Button index 11 : SW sidebar switch
-	bool SuperVisible { true };
-
-	// TODO New buttons
-
-	// Button index 61-68 : Heros buttons
+	// Button index 1-8 : Heros buttons
 	bool HeroVisible { true };
+	const wchar_t* HoveredHero { nullptr };
 
-	// Button index 71-100 : Select buttons
+	// Button index 11-100 : Select buttons
 	bool UpdateSelect { false };
-	int RecordIndex { 71 };
+	int RecordIndex { 11 };
 	std::vector<SelectRecordStruct> CurrentSelectCameo {};
 	const wchar_t* HoveredSelected { nullptr };
+
+	// TODO New buttons
 
 private:
 	int ButtonIndex { -1 }; // -1 -> above no buttons, 0 -> above buttons background, POSITIVE -> above button who have this index